requirements.txt
certifi~=2023.7.22
motor~=3.3.1
beanie==1.23.1
fastapi==0.104.0
pydantic~=2.4.2
pymongo~=4.5.0
uvicorn~=0.23.2
pyJWT~=2.8.0
python-decouple==3.8
decouple~=0.0.7
starlette==0.31.1
email-validator==2.1.0.post1
cryptography
requests~=2.31.0
beautifulsoup4~=4.12.2